+++ src/main/java/br/com/rentacarserver/rentacarserver/services/CustomersService.java
@@ -1,6 +1,7 @@
 package br.com.rentacarserver.rentacarserver.services;
 
 import br.com.rentacarserver.rentacarserver.entities.CustomersEntity;
+import br.com.rentacarserver.rentacarserver.myutils.MyUtils;
 import br.com.rentacarserver.rentacarserver.repositories.CustomersRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -25,21 +26,42 @@ public class CustomersService {
         return this.customersRepository.findByName(customersName);
     }
 
-    public void insertBrand(@RequestBody final CustomersEntity customersEntity){
-        try{
-            this.customersRepository.save(customersEntity);
-        }catch (Exception e){
-            e.getMessage();
+    public void insertCustomers(@RequestBody final CustomersEntity customersEntity){
+        if(customersEntity.getName() != null &&
+                (!customersEntity.getName().isEmpty())){
+            if(customersEntity.getId() == null ||
+                    (!this.customersRepository.existsById(customersEntity.getId()))){
+                customersEntity.setCreatedAt(MyUtils.timestamp);
+                this.customersRepository.save(customersEntity);
+            }else {
+                throw new IllegalArgumentException("Invalid ID");
+            }
+        }else{
+            throw new IllegalArgumentException("Name as null or empty");
         }
     }
 
     public void updateCustomers(@RequestBody final CustomersEntity customersEntity){
-        this.customersRepository.save(customersEntity);
+        if(customersEntity.getId() != null &&
+                (this.customersRepository.existsById(customersEntity.getId()))) {
+            if(customersEntity.getName() != null &&
+                    (!customersEntity.getName().isEmpty())) {
+                customersEntity.setUpdateAt(MyUtils.timestamp);
+                this.customersRepository.save(customersEntity);
+            }else{
+                throw new IllegalArgumentException("Name can't be null");
+            }
+        }else{
+            throw new IllegalArgumentException("Update need a valid id");
+        }
     }
 
     public void deleteCustomersById(@PathVariable(value = "id") Long customersId){
-        if(customersId != null) {
+        if(customersId != null &&
+                (this.customersRepository.existsById(customersId))){
             this.customersRepository.deleteById(customersId);
+        }else{
+            throw new IllegalArgumentException("id can't be null");
         }
     }
